+++ changelog.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
-## [Unreleased]
+## [2.0.0] - 2024/08/13
 
 ### Added
 
@@ -70,6 +70,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Init webpack build
 - Init linter
 
+[2.0.0]: https://github.com/ditrit/leto-modelizer-plugin-cli/blob/2.0.0/changelog.md
 [1.0.0]: https://github.com/ditrit/leto-modelizer-plugin-cli/blob/1.0.0/changelog.md
 [0.2.1]: https://github.com/ditrit/leto-modelizer-plugin-cli/blob/0.2.1/changelog.md
 [0.2.0]: https://github.com/ditrit/leto-modelizer-plugin-cli/blob/0.2.0/changelog.md